Moving file to api-objects where it should be
